refactor: update todolist
