+++ src/composables/firstpage-functions.ts
@@ -1,13 +1,13 @@
 import { ref, computed } from 'vue'
 
-export const isGay = ref(false)
+export const isGray = ref(false)
 
-export function toggleSlayColor() {
-  isGay.value = !isGay.value
+export function toggleColor() {
+  isGray.value = !isGray.value
 }
 
-export const sassify = computed(() => {
-  return isGay.value
+export const changeColor = computed(() => {
+  return isGray.value
     ? { background: 'linear-gradient(135deg, #153677, #4e085f)' }
     : { background: '#424549' }
 })
