+++ src/views/todo-page.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
-import TodoPage from '../components/todopage-components.vue'
+import TodoPage from '../components/todo-page.vue'
 </script>
 
 <template>
   <div>
-    <TodoPage pageHeader="To-Do List" toHome="<" />
+    <TodoPage pageHeader="To-Do List" />
   </div>
 </template>
